requirements.txt
asgiref==3.7.2
Django==4.2.7
django-axes==6.1.1
Pillow==10.1.0
pylibmc==1.6.3
python-memcached==1.59
shell==1.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
